# backend/api-gateway/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Sistema de Delivery - API Gateway",
    "description": "Documentación centralizada de todos los microservicios que componen el Sistema de Delivery. A través de esta interfaz unificada se puede interactuar y probar los endpoints de cada servicio.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5000",
      "description": "API Gateway Local"
    }
  ],
  "paths": {
    "/api/usuarios": {
      "get": {
        "tags": ["Usuarios"],
        "summary": "Listar todos los usuarios",
        "responses": {
          "200": {
            "description": "Lista de usuarios obtenida exitosamente"
          }
        }
      },
      "post": {
        "tags": ["Usuarios"],
        "summary": "Crear un nuevo usuario",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "email": { "type": "string" },
                  "telefono": { "type": "string" },
                  "rol": { "type": "string", "enum": ["cliente", "administrador", "soporte"] }
                },
                "required": ["nombre", "email", "telefono"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario creado exitosamente"
          }
        }
      }
    },
...
    "/api/sucursales": {
      "get": {
        "tags": ["Sucursales"],
        "summary": "Listar todas las sucursales",
        "responses": {
          "200": {
            "description": "Lista de sucursales"
          }
        }
      },
      "post": {
        "tags": ["Sucursales"],
        "summary": "Crear una nueva sucursal",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "direccion": { "type": "string" },
                  "latitud": { "type": "number" },
                  "longitud": { "type": "number" }
                },
                "required": ["nombre", "direccion", "latitud", "longitud"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucursal creada con éxito"
          }
        }
      }
    },
...
    "/api/inventario": {
      "get": {
        "tags": ["Inventario / Productos"],
        "summary": "Listar todos los productos",
        "responses": {
          "200": {
            "description": "Lista de productos"
          }
        }
      },
      "post": {
        "tags": ["Inventario / Productos"],
        "summary": "Crear un producto en el catálogo maestro",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "descripcion": { "type": "string" },
                  "precio": { "type": "number" },
                  "categoria": { "type": "string" }
                },
                "required": ["nombre", "precio"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado con éxito"
          }
        }
      }
    },
...
    "/api/pedidos": {
      "get": {
        "tags": ["Pedidos"],
        "summary": "Listar todos los pedidos",
        "responses": {
          "200": {
            "description": "Lista de pedidos"
          }
        }
      },
      "post": {
        "tags": ["Pedidos"],
        "summary": "Crear un nuevo pedido",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "clienteId": { "type": "string" },
                  "sucursalId": { "type": "string" },
                  "productos": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "productoId": { "type": "string" },
                        "cantidad": { "type": "number" },
                        "precioUnitario": { "type": "number" }
                      }
                    }
                  },
                  "total": { "type": "number" },
                  "direccionEntrega": { "type": "string" },
                  "metodoPago": { "type": "string", "enum": ["efectivo", "tarjeta", "transferencia"] }
                },
                "required": ["clienteId", "sucursalId", "productos", "total", "direccionEntrega"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido creado con éxito"
          }
        }
      }
    },
...
    "/api/repartidores": {
      "get": {
        "tags": ["Repartidores"],
        "summary": "Listar todos los repartidores",
        "responses": {
          "200": {
            "description": "Lista de repartidores"
          }
        }
      },
      "post": {
        "tags": ["Repartidores"],
        "summary": "Crear un nuevo repartidor",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "nombre": { "type": "string" },
                  "telefono": { "type": "string" },
                  "vehiculo": { "type": "string" },
                  "estado": { "type": "string", "enum": ["activo", "inactivo", "en_entrega"] }
                },
                "required": ["nombre", "telefono", "vehiculo"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Repartidor creado con éxito"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}